refactor: Update NumberInput component to include min and max attributes
The NumberInput component in the atoms folder has been updated to include the min and max attributes. This allows for setting the minimum and maximum values that can be entered in the input field. The default values for min and max are 0 and 1000000 respectively. This change ensures better control over the input range for number values.