Refactor HttpUtils class: Add method to abort proposal
This commit refactors the HttpUtils class by adding a new method `abortProposal` that allows for the abortion of a proposal. The method sends an HTTP DELETE request to the specified URL with the proposal ID as a parameter. It also sets the `user` header to "supplier-engagement-bot". If the response code is not 200, an IOException is thrown.

The `abortProposal` method is intended to be used for auto-expiring engagements after 48 hours and sending a warning email. This functionality was added as part of a recent user commit.

Other changes in this commit include organizing imports, adding log messages, and updating the logging configuration.